mwpw-155030 - authorable pagination type
