+++ eds/blocks/partner-news/partner-news.js
@@ -32,10 +32,15 @@ export default async function init(el) {
     '{{filters}}': 'Filters',
     '{{last-90-days}}': 'Last 90 days',
     '{{load-more}}': 'Load more',
+    '{{next}}': 'Next',
+    '{{next-page}}': 'Next Page',
     '{{no-results-description}}': 'Try checking your spelling or broadening your search.',
     '{{no-results-title}}': 'No Results Found',
     '{{of}}': 'Of',
+    '{{page}}': 'Page',
+    '{{prev}}': 'Prev',
     '{{previous-month}}': 'Previous month',
+    '{{previous-page}}': 'Previous Page',
     '{{results}}': 'Results',
     '{{search}}': 'Search',
     '{{show-all}}': 'Show all',
@@ -72,6 +77,7 @@ export default async function init(el) {
     cardsPerPage: 12,
     ietf: config.locale.ietf,
     collectionTags: '"caas:adobe-partners/collections/news"',
+    pagination: 'load-more',
   };
 
   const app = document.createElement('partner-news');
