+++ eds/components/PartnerCardsStyles.js
@@ -537,66 +537,7 @@ export const partnerCardsStyles = css`
   .pagination-wrapper > * {
     padding: 5px 0;
   }
-  
-  .pagination-wrapper .pagination-pages-list button {
-    border: none;
-    background-color: transparent;
-  }
-  
-  .pagination-wrapper .page-btn {
-    position: relative;
-    min-width: 32px;
-    min-height: 32px;
-    max-width: 70px;
-    padding: 8px 10px;
-    font-size: .875rem;
-    line-height: .875rem;
-    color: #4b4b4b;
-    word-break: break-word;
-    text-align: center;
-    white-space: nowrap;
-    text-overflow: ellipsis;
-    border-radius: 16px;
-    overflow: hidden;
-  }
-  
-  .pagination-wrapper .pagination-prev-btn {
-    padding-right: 10px;
-    font-weight: 700;
-  }
-  
-  .pagination-wrapper .pagination-next-btn {
-    padding-left: 10px;
-    font-weight: 700;
-  }
-  
-  .pagination-wrapper .pagination-prev-btn:hover,
-  .pagination-wrapper .pagination-next-btn:hover,
-  .pagination-wrapper .page-btn:hover {
-    cursor: pointer;
-    text-decoration: underline;
-    transition: all .3s;
-  }
-  
-  .pagination-wrapper .pagination-prev-btn.disabled,
-  .pagination-wrapper .pagination-prev-btn.disabled:hover,
-  .pagination-wrapper .pagination-next-btn.disabled,
-  .pagination-wrapper .pagination-next-btn.disabled:hover,
-  .pagination-wrapper .page-btn.selected:hover {
-    text-decoration: none;
-    cursor: default;
-  }
-  
-  .pagination-wrapper .pagination-prev-btn.disabled,
-  .pagination-wrapper .pagination-next-btn.disabled {
-    color: #9d9d9d;
-  }
-  
-  .pagination-wrapper .page-btn.selected {
-    background-color: #e5e5e5;
-    border-radius: 50%;
-  }
-  
+
   .pagination-wrapper .pagination-total-results {
     font-weight: 700;
     text-transform: lowercase;
@@ -897,148 +838,13 @@ export const partnerCardsStyles = css`
   }
 `;
 
-export const newsCardStyles = css`
-
-  .news-card * {
-    box-sizing: border-box;
-  }
-  
-  .news-card {
-    display: flex;
-    flex-direction: column;
-    background-color: #fff;
-    border: 1px solid ${borderColor};
-    border-radius: 4px;
-    overflow: hidden;
-    width: 100%;
-    height: 400px;
-  }
-  
-  .news-card:hover {
-    box-shadow: 0 3px 6px 0 rgba(0,0,0,.16);
-    transition: box-shadow .3s ease-in-out;
-  }
-  
-  .news-card .card-header {
-    min-height: 192px;
-    max-height: 192px;
-    background-color: #323232;
-    background-repeat: no-repeat;
-    background-position: 50% 50%;
-    background-size: cover;
-    border-top-left-radius: 4px;
-    border-top-right-radius: 4px;
-    position: relative;
-  }
-  
-  .news-card:hover .card-header:after {
-    opacity: 1;
-  }
-  
-  .news-card .card-header:after {
-    position: absolute;
-    content: "";
-    top: 0;
-    bottom: 0;
-    left: 0;
-    right: 0;
-    margin: auto;
-    z-index: 0;
-    background-color: rgba(0,0,0,.35);
-    opacity: 0;
-    transition: opacity .3s ease-in-out;
-  }
-  
-  .news-card .card-content {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    background-color: #fff;
-    padding: 16px 16px 20px;
-    height: 100%;
-  }
-  
-  .news-card .card-title {
-    color: #323232;
-    font-size: 1.125rem;
-    font-weight: 700;
-    line-height: 1.375rem;
-    word-break: break-word;
-    max-height: 2.75rem;
-    margin: 0 0 7px;
-    display: -webkit-box;
-    -webkit-box-orient: vertical;
-    overflow: hidden;
-    -webkit-line-clamp: 2;
-  }
-  
-  .news-card .card-description {
-    color: #505050;
-    font-size: .875rem;
-    line-height: 1.3125rem;
-    font-weight: 400;
-    word-break: break-word;
-    margin: 0 0 14px;
-    max-height: 3.9375rem;
-    display: -webkit-box;
-    -webkit-box-orient: vertical;
-    overflow: hidden;
-    -webkit-line-clamp: 3;
-  }
-  
-  .news-card .card-footer {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-  }
-  
-  .news-card .card-btn {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-size: .875rem;
-    line-height: 1.063rem;
-    font-weight: 700;
-    min-width: 60px;
-    max-width: 100%;
-    height: 28px;
-    max-height: 32px;
-    margin-left: 16px;
-    padding-left: 14px;
-    padding-right: 14px;
-    text-decoration: none;
-    color: #222222;
-    border: 2px solid #222222;
-    border-radius: 16px;
-    background-color: #ffffff00;
-    cursor: pointer;
-    transition: border-color .3s ease-in-out,background-color .3s ease-in-out;
-  }
-  
-  .news-card .card-btn:hover {
-    text-decoration: none;
-    border-color: #222222;
-    background-color: #222222;
-    color: #ffffff;
-  }
-  
-  .news-card .card-date {
-    color: #747474;
-    font-size: .875rem;
-    line-height: 1.3125rem;
-    word-break: break-word;
-    max-height: 3.9375rem;
-    padding: 0;
-  }
-`;
-
-export const loadMorePaginationStyles = css`
-  .pagination-wrapper {
+export const partnerCardsLoadMoreStyles = css`
+  .pagination-wrapper.pagination-wrapper-load-more {
     justify-content: center;
     flex-direction: column-reverse;
   }
   
-  .pagination-wrapper .load-more-btn {
+  .pagination-wrapper-load-more .load-more-btn {
     display: flex;
     justify-content: center;
     align-items: center;
@@ -1059,7 +865,7 @@ export const loadMorePaginationStyles = css`
     transition: border-color .3s ease-in-out,background-color .3s ease-in-out;
   }
   
-  .pagination-wrapper .load-more-btn:hover {
+  .pagination-wrapper-load-more .load-more-btn:hover {
     text-decoration: none;
     border-color: #222222;
     background-color: #222222;
@@ -1067,20 +873,79 @@ export const loadMorePaginationStyles = css`
   }
 `;
 
-export const numericPaginationStyles = css`
-  .pagination-wrapper {
+export const partnerCardsPaginationStyles = css`
+  .pagination-wrapper.pagination-wrapper-default {
     flex-wrap: wrap;
     justify-content: space-between;
   }
   
   @media screen and (max-width: 1200px) {
-    .pagination-wrapper {
+    .pagination-wrapper.pagination-wrapper-default {
       justify-content: center;
     }
   }
+  
+  .pagination-wrapper-default .pagination-pages-list button {
+    border: none;
+    background-color: transparent;
+  }
+  
+  .pagination-wrapper-default .page-btn {
+    position: relative;
+    min-width: 32px;
+    min-height: 32px;
+    max-width: 70px;
+    padding: 8px 10px;
+    font-size: .875rem;
+    line-height: .875rem;
+    color: #4b4b4b;
+    word-break: break-word;
+    text-align: center;
+    white-space: nowrap;
+    text-overflow: ellipsis;
+    border-radius: 16px;
+    overflow: hidden;
+  }
+  
+  .pagination-wrapper-default .pagination-prev-btn {
+    padding-right: 10px;
+    font-weight: 700;
+  }
+  
+  .pagination-wrapper-default .pagination-next-btn {
+    padding-left: 10px;
+    font-weight: 700;
+  }
+  
+  .pagination-wrapper-default .pagination-prev-btn:hover,
+  .pagination-wrapper-default .pagination-next-btn:hover,
+  .pagination-wrapper-default .page-btn:hover {
+    cursor: pointer;
+    text-decoration: underline;
+    transition: all .3s;
+  }
+  
+  .pagination-wrapper-default .pagination-prev-btn.disabled,
+  .pagination-wrapper-default .pagination-prev-btn.disabled:hover,
+  .pagination-wrapper-default .pagination-next-btn.disabled,
+  .pagination-wrapper-default .pagination-next-btn.disabled:hover,
+  .pagination-wrapper-default .page-btn.selected:hover {
+    text-decoration: none;
+    cursor: default;
+  }
+  
+  .pagination-wrapper-default .pagination-prev-btn.disabled,
+  .pagination-wrapper-default .pagination-next-btn.disabled {
+    color: #9d9d9d;
+  }
+  
+  .pagination-wrapper-default .page-btn.selected {
+    background-color: #e5e5e5;
+    border-radius: 50%;
+  }
 `;
 
-export const dateFilterStyles = css`
+export const partnerCardsDateFilterStyles = css`
   .date-filter-tag {
     background-color: transparent;
     border: none;
